Adds support for logging lambdas in nitro::log
This allows to only evaluate expensive expressions in log lines when
the log is actually written. If for some reason the log gets filtered,
an expensive expression within a lambda won't be evaluated.

Note:
For performance reasons, make sure that the creation of the lambda,
especially the binding of captured variables is cheap.

Note:
If you pass a class, which has an `operator()` defined with the
signature `std::string()` and an overloaded stream `operator<<`, the
`operator()` takes precedence when evaluating the whole expression.

For example, the following code will log "A":
```
struct Foo
{
  std::string operator(){ return "A"; }
};

std::ostream& operator<<(std::ostream& s, Foo& f)
{
  s << "B";
}

log::info() << Foo();
```