feat(options): Adds greedy position parsing
This allows to interpret all passed arguments after the first detected positional argument as positional arguments as well. In particular, this is useful for prefix-commands, as the executed command is usually the first positional and every argument passed after that belongs to the executed command and not the prefix command.